requirements.txt
Django==4.0.2
Pillow==9.0.1
django-admin-sortable2==1.0.4
